refactor: 추모관 페이지 하나로 통일
