+++ src/Router.jsx
@@ -12,8 +12,6 @@ import { HomePage } from "./pages/HomePage";
 
 /* 추모관 관련 */
 import MemorialHallPage from "./pages/MemorialHallPage";
-import MemorialMyHallPage from "./pages/MemorialMyHallPage";
-import { MemorialManagerHallPage } from "./pages/MemorialManagerHallPage";
 import { ProfileEditPage } from "./features/MemorialHall/pages/ProfileEditPage";
 import OpenMemorialHallPage from "./pages/OpenMemorialHallPage";
 import { EnterMemorialHallPage } from "./pages/EnterMemorialHallPage";
@@ -51,12 +49,12 @@ const router = createBrowserRouter([
       { path: "/memorial", element: <MemorialHallPage />, loader: authGuard },
       {
         path: "/memorial-my",
-        element: <MemorialMyHallPage />,
+        element: <MemorialHallPage />,
         loader: authGuard,
       },
       {
         path: "/memorial-manager",
-        element: <MemorialManagerHallPage />,
+        element: <MemorialHallPage />,
         loader: authGuard,
       },
       {
